Create code_pygame
The two cars are there one is the thief car and the other is the police car. The  police car helps to find the thief car. Initially the police car speed is made greater than the thief car by the programmer so that the thief is caught easily. This will work in the form of game named car game. The player can control both the car by their own.